gh-923 Fixed a bug in ParquetStore which allowed the index to be unsorted

+++ store-implementation/parquet-store/src/test/java/uk/gov/gchq/gaffer/parquetstore/operation/EdgeCasesTest.java
@@ -74,12 +74,17 @@ public class EdgeCasesTest {
     @AfterClass
     public static void cleanUp() throws IOException {
         final FileSystem fs = FileSystem.get(new Configuration());
-        final ParquetStoreProperties props = new ParquetStoreProperties();
-        Path dataDir = new Path(props.getDataDir());
-        fs.delete(dataDir, true);
-        while (fs.listStatus(dataDir.getParent()).length == 0) {
-            dataDir = dataDir.getParent();
+        deleteFolder("parquet_data", fs);
+    }
+
+    private static void deleteFolder(final String path, final FileSystem fs) throws IOException {
+        Path dataDir = new Path(path);
+        if (fs.exists(dataDir)) {
             fs.delete(dataDir, true);
+            while (fs.listStatus(dataDir.getParent()).length == 0) {
+                dataDir = dataDir.getParent();
+                fs.delete(dataDir, true);
+            }
         }
     }
 
@@ -126,6 +131,7 @@ public class EdgeCasesTest {
                 EdgeCasesTest.class.getResourceAsStream("/schemaUsingStringVertexType/storeSchema.json"),
                 EdgeCasesTest.class.getResourceAsStream("/schemaUsingStringVertexType/storeTypes.json"));
         ParquetStoreProperties parquetProperties = getParquetStoreProperties();
+        parquetProperties.setDataDir("readElementsWithZeroElementFiles");
         parquetProperties.setAddElementsOutputFilesPerGroup(200);
         parquetProperties.setAddElementsBatchSize(1024);
         ParquetStore store = new ParquetStore(gafferSchema, parquetProperties);
@@ -138,40 +144,13 @@ public class EdgeCasesTest {
         final Iterable<? extends Element> retrievedElements = graph.execute(new GetAllElements(), USER);
         final Iterator<? extends Element> iter = retrievedElements.iterator();
         assertTrue(iter.hasNext());
-        Edge edge = (Edge) iter.next();
-        assertEquals("BasicEdge", edge.getGroup());
-        assertEquals("src0", edge.getSource());
-        assertEquals("dst0", edge.getDestination());
-        assertEquals(false, edge.isDirected());
-        assertEquals(2, (int) edge.getProperty("count"));
-        edge = (Edge) iter.next();
-        assertEquals("src0", edge.getSource());
-        assertEquals("dst0", edge.getDestination());
-        assertEquals(true, edge.isDirected());
-        assertEquals(2, (int) edge.getProperty("count"));
-
-        for (int i = 0; i <98 ; i++) {
+        int counter = 0;
+        while (iter.hasNext()) {
             iter.next();
+            counter++;
         }
-
-        Entity e = (Entity) iter.next();
-        assertEquals("BasicEntity", e.getGroup());
-        assertEquals("vert0", e.getVertex());
-        assertEquals(2, (int) e.getProperty("count"));
-        e = (Entity) iter.next();
-        assertEquals("BasicEntity", e.getGroup());
-        assertEquals("vert1", e.getVertex());
-        assertEquals(2, (int) e.getProperty("count"));
-
-        for (int i = 0; i <48 ; i++) {
-            iter.next();
-        }
-
-        e = (Entity) iter.next();
-        assertEquals("BasicEntity2", e.getGroup());
-        assertEquals("vert9", e.getVertex());
-        assertEquals(2, (int) e.getProperty("count"));
-        assertFalse(iter.hasNext());
+        assertEquals(150, counter);
+        deleteFolder("readElementsWithZeroElementFiles", FileSystem.get(new Configuration()));
     }
 
     @Test
